make TelegramID field int(it was int64)

+++ model/user.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type User struct {
 	gorm.Model
 	ChatID     int64 `gorm:"index"`
-	TelegramID int64 `gorm:"index"`
+	TelegramID int   `gorm:"index"`
 	State      UserState
 	ISU        uint64
 }
